fix: emit new event specifically when all data has been written to disk (#8)
The `finish` event on the target Writable itself would not indicate
this when gzip compression is being used, because the gzip stream
could finish before the subsequent `fs.write` call to the actual
target file was completed. This could truncate the file in cases
where the application quits immediately after the 'finish' event
was emitted.